% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_line_breaks.R
\name{add_line_breaks}
\alias{add_line_breaks}
\title{Split long sentences into multiple lines by adding line breaks}
\usage{
add_line_breaks(string, words)
}
\arguments{
\item{string}{Pass text string here.}

\item{words}{The maximum number of words allowed in the first line}
}
\description{
Split long sentences into multiple lines by adding line breaks
}
\examples{
library(stringr)
library(magrittr)
add_line_breaks("This is a lovely cup of tea",2)
add_line_breaks("This is a lovely cup of tea",3)
add_line_breaks("This is a lovely cup of tea",1)
add_line_breaks("Oh wow. Amazing. I cannot believe that this is actually possible!",4)
paste0("word",1:100, collapse = " ") \%>\% add_line_breaks(3)
}
