% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakitdown.r
\name{breakitdown}
\alias{breakitdown}
\title{Contingency table by referencing data and variables}
\usage{
breakitdown(x, var_org, var_per)
}
\arguments{
\item{x}{Data frame to be passed through}

\item{var_org}{The organising variable, i.e. variable for grouping categories. \cr
For instance, age could be the grouping variable if you want to compare frequencies across age-groups.}

\item{var_per}{The variable for which you want the percentages / frequencies.}
}
\description{
Produce a contingency table with percentage signs.
}
\examples{
breakitdown(mtcars,am,gear)
# A tibble: 4 x 4
# Groups:   am [2]
     am  gear     n rel.freq
* <dbl> <dbl> <int>    <chr>
1     0     3    15      79\%
2     0     4     4      21\%
3     1     4     8      62\%
4     1     5     5      38\%
}
\keyword{contingency}
\keyword{table}
