% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel_clean.R
\name{camel_clean}
\alias{camel_clean}
\title{Convert "CamelCase" to "Camel Case"}
\usage{
camel_clean(text)
}
\arguments{
\item{text}{character vector in 'CamelCase' format to format}
}
\value{
Returns a formatted string.
}
\description{
Convert a text string from the format "CamelCase" to "Camel Case". This is
used for converting variable names such as "ApplePie" to "Apple Pie" for the
purpose of prettifying plot labels.
}
\examples{
camel_clean("NoteHowTheStringIsFormatted")

}
