% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_first.R
\name{cap_first}
\alias{cap_first}
\title{Capitalise the first letter of all words}
\usage{
cap_first(string)
}
\arguments{
\item{string}{String or character vector to pass through.}
}
\description{
Capitalise the first letter of all words
}
\examples{
cap_first("steeles pots and pans")

}
