% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.R
\name{comma}
\alias{comma}
\title{Add comma separator for thousands}
\usage{
comma(x)
}
\arguments{
\item{x}{numeric value to add thousand separator.}
}
\value{
Returns a formatted string.
}
\description{
Takes a numeric value and returns a character value
which is rounded to the whole number, and adds a comma
separator at the thousands. A convenient wrapper function
around \code{scales::label_comma()}.
}
\examples{
comma(1000 * 1000)

comma(20190721)

}
