% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ngram.R
\name{count_ngram}
\alias{count_ngram}
\title{Summarise the counts of a specified ngram in character vector}
\usage{
count_ngram(text, count_col = "count", n = 2)
}
\arguments{
\item{text}{character. Vector containing text to split into n-grams and count.}

\item{count_col}{A string vector for the name of the count column. Defaults to "count"}

\item{n}{Specify n of ngram}
}
\value{
data frame with the count results, containing two columns:
  - `text`: n-grams identified
  - `count` (unless otherwise specified)
}
\description{
Summarise the counts of a specified ngram in character vector
}
\examples{
count_ngram(c("The quick brown fox jumped over the lazy dog"))

}
