% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab mc.r
\name{ctab.mc}
\alias{ctab.mc}
\title{Contingency table for a multi-code question}
\usage{
ctab.mc(x)
}
\arguments{
\item{vec}{Data frame to be passed through.
Use select() to remove any unnecessary columns (including ID).}
}
\description{
Use a data frame as argument, and a contingency table by column percentages is returned as output.
First column will display the codes for the question
Each subsequent column of the table will show the column percentages for each question code.
}
\examples{
Coming soon
}
\keyword{contingency}
\keyword{table}
