% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likert scale reverser.R
\name{likert.reverse}
\alias{likert.reverse}
\title{Reverse a Likert scale}
\usage{
likert.reverse(x, top, bottom)
}
\arguments{
\item{x}{Numeric vector}

\item{top}{Top value of the scale for the variable, e.g. 10 for a 0-10 scale.}

\item{top}{Bottom value of the scale for the variable, e.g. 0 for a 0-10 scale.}
}
\description{
Reverse a Likert scale such that on a 0-10 scale, 10 becomes 0 and 0 becomes 10.
}
\examples{
data <-c(5, 4, 3, 2, 6)
likert.reverse(data, 6, 0)
[1] 1 2 3 4 0
}
\keyword{likert}
\keyword{scale}
