% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwords.R
\name{nwords}
\alias{nwords}
\title{Count the number of words in text string}
\usage{
nwords(string, pseudo = FALSE)
}
\arguments{
\item{string}{Pass text string here.}

\item{pseudo}{Determines whether groups of special characters are matched. Defaults to FALSE (not matched)}
}
\description{
Count the number of words in text string
}
\examples{
nwords("Oh my what a lovely day. We should all go out and play!")

}
