% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_arrange.R
\name{str_arrange}
\alias{str_arrange}
\title{Sorts letters in a character string by alphabetical order}
\usage{
str_arrange(text, decreasing = FALSE)
}
\arguments{
\item{text}{A vector of character string to pass through.}

\item{decreasing}{logical, specifies whether sort is increasing or decreasing.
See \code{sort()} (base)}
}
\value{
Returns a formatted string.
}
\description{
Sorts letters in a character string by alphabetical order
}
\examples{
str <- c("sugar", "spice", "everything nice")
str_arrange(text = str)

}
