% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_left.R
\name{str_left}
\alias{str_left}
\title{Return the n number of characters from the left}
\usage{
str_left(x, n)
}
\arguments{
\item{x}{String character to pass through}

\item{n}{Integer specifying the number of characters to return from the left.}
}
\description{
Wrapper convenience function
Analagous to `str_right()`
}
\examples{
str_left("Wittgenstein", 7)

}
