% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace.R
\name{str_replace}
\alias{str_replace}
\title{Replace matched patterns in a string}
\usage{
str_replace(string, pattern, replacement)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern}{Pattern to look for.
A character vector of replacements. Should be either length one, or the same length as string or pattern.}

\item{replacement}{}
}
\description{
An implementation of `stringr::str_replace()` using only base R.
Vectorised over `string`, `pattern`, and `replacement`
}
