% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_from_match.R
\name{subset_from_match}
\alias{subset_from_match}
\title{Subset a List from a Matched String}
\usage{
subset_from_match(x, key, before = FALSE)
}
\arguments{
\item{x}{A character vector to perform the matching and subsetting on.}

\item{key}{A character string to match in \code{x}.}

\item{before}{A logical value indicating whether the strings \emph{before} the
matched key should be returned. Defaults to FALSE. If set to TRUE, strings
\emph{before} the matched key are returned.}
}
\value{
By default, a character vector starting from the first occurrence of
\code{key} in \code{x} and including all subsequent elements. Logic can be reversed
with \code{before} argument. If \code{key} is not found in \code{x}, the function returns
\code{NULL}.
}
\description{
This function takes a list of strings and a key string as input.
It returns a new list that starts from the first occurrence of the key
string in the original list and includes all the strings that follow it. If
the key string is not found in the list, the function returns \code{NULL}.
Optional argument \code{before} allows the logic to be reversed, returning a new
list that start from the first value in the original list up to the matched
string.
}
\examples{
x <- c("a", "b", "c", "d", "e")
key <- "c"
subset_from_match(x, key)  # Returns: "c" "d" "e"

}
