% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans.m}
\alias{trans.m}
\title{Translate function using the Microsoft Translator API}
\usage{
trans.m(what, from = NULL, to = "en", key = MyKey)
}
\arguments{
\item{what}{What text string to translate}

\item{from}{What language to translate from. Defaults to NULL}

\item{to}{What language to translate to. Defaults to "en"}
}
\description{
This function allows you to translate character strings through the Microsoft Translator API.
To use this function, you will need to first acquire an access token and pass it through the key argument.
}
\examples{
trans.m(what="Blanc",from="fr",to="en",key="F023kljadfoilkjlkj")
}
\keyword{translate}
