% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans.t}
\alias{trans.t}
\title{Translate function using transltr.org API}
\usage{
trans.t(what = NULL, from = NULL, to = "en")
}
\arguments{
\item{what}{What text string to translate}

\item{from}{What language to translate from. Defaults to NULL}

\item{to}{What language to translate to. Defaults to "en"}
}
\description{
This function allows you to translate character strings through the transltr.org API.
}
\examples{
trans.t(what="Blanc",from="fr",to="en")
}
\keyword{translate}
