% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_transltr.R
\name{trans.transltr}
\alias{trans.transltr}
\title{Translate function using transltr.org API}
\usage{
trans.transltr(what, from = NULL, to)
}
\arguments{
\item{what}{What tex string to translate}

\item{from}{What language to translate from. Defaults to NULL}

\item{to}{What language to translate to? Mandatory}
}
\description{
This function allows you to translate character strings through the transltr.org API.
}
\examples{
trans.transltr()
}
\keyword{translate}
