% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Add a character at the start and end of a character string}
\usage{
wrap(text, wrapper = "\\"")
}
\arguments{
\item{text}{string to be add 'wrapping' characters.}

\item{wrapper}{string to be added around value supplied in `text`.}
}
\value{
Returns a formatted string.
}
\description{
Add a character at the start and end of a
  character string, where the default behaviour is to add a double quote.
}
\examples{
wrap(text = "lol")

wrap(text = "lol", wrapper = "lol")

}
