% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_line_breaks.R
\name{add_line_breaks}
\alias{add_line_breaks}
\title{Split long strings into multiple lines by adding line breaks}
\usage{
add_line_breaks(string, nword = NULL, nchar = NULL)
}
\arguments{
\item{string}{Pass text string here.}

\item{nword}{numeric, defaults to NULL. The maximum number of words allowed in the first line. Only one
argument from either `nword` or `nchar` is used, otherwise an error is
generated.}

\item{nchar}{numeric, defaults to NULL. Number of character units by which the
next space would be replaced with `\n` to move text to next line. Only one
argument from either `nword` or `nchar` is used, otherwise an error is
generated.}
}
\description{
Wrap text in visualizations according to a preset character
  threshold, or the maximum number of words allowed in the first line. The
  next space in the string is replaced with `\n`, which will render as next
  line in plots and messages.
}
\examples{
tea_text <- "This is a lovely cup of tea"

add_line_breaks(string = tea_text, nword = 2)

add_line_breaks(string = tea_text, nword = 3)

# Display text in a plot
plot.new()
text(
  x = 0.5,
  y = 0.5,
  labels = add_line_breaks(string = tea_text, nword = 3),
  cex = 1.6
)

# Applying to a long string
paste0("word",1:100, collapse = " ") \%>\%
  add_line_breaks(nword = 3) \%>\%
  message()

# Define breaks using characters
message(
  add_line_breaks(
    "Oh wow. Amazing. I cannot believe that this is actually possible!",
    nchar = 10)
  )

}
