% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.r
\name{ctab}
\alias{ctab}
\title{Pass a factor variable and return a contingency table by column percentages}
\usage{
ctab(vec)
}
\arguments{
\item{vec}{Vector to be passed through. Must be a factor variable}
}
\description{
Pass a factor variable and return a contingency table by column percentages
}
\examples{
vec <- sample(c("Sandwich","Porridge","Muesli/cereal","Other","No breakfast"),2000,replace=TRUE,prob=c(0.2,0.15,0.25,0.15,.25))
vec <- factor(vec)
ctab(vec)
}
\keyword{contingency}
\keyword{table}
