% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwords.R
\name{nwords}
\alias{nwords}
\title{Count the number of words in text string}
\usage{
nwords(text, pseudo = FALSE)
}
\arguments{
\item{text}{Pass text string here.}

\item{pseudo}{Determines whether groups of special characters are matched.
Defaults to `FALSE` (not matched)}
}
\value{
numeric value containing count of words.
}
\description{
Count the number of words in text string
}
\examples{
nwords(text = "Oh my what a lovely day. We should all go out and play!")

}
