% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2hex.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert RGB values to HEX colour codes}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r}{Value for r}

\item{g}{Value for g}

\item{b}{Value for b}
}
\description{
Suited for using in viz packages like {wordcloud2}
A convenience wrapper around `rgb()`
}
\examples{
rgb2hex(0,144,218)

}
