% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_left.R
\name{str_left}
\alias{str_left}
\title{Return the `n` number of characters from the left}
\usage{
str_left(text, n)
}
\arguments{
\item{text}{String character to pass through}

\item{n}{Integer specifying the number of characters to return from the left.}
}
\value{
Returns a formatted string.
}
\description{
This is a convenience wrapper function around `str_sub()`, and is analagous
to `str_right()`.
}
\examples{
str_left("Wittgenstein", 7)

}
