% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_nth.R
\name{str_replace_nth}
\alias{str_replace_nth}
\title{Replace nth occurring pattern in a string}
\usage{
str_replace_nth(x, pattern, replacement, n)
}
\arguments{
\item{x}{Pass string character}

\item{pattern}{String containing characters to match}

\item{replacement}{String containing characters to replace.}

\item{n}{Nth term to be replaced}
}
\description{
Replaces the nth occurrence of a specified pattern in a string. This enables
a more targetted way of replacing patterns compared to
`stringr::str_replace()` and `stringr::str_replace_all()`.
}
\examples{
x <- "pineapplepie"

str_replace_nth(x = x, pattern = "p", replacement = "q", n = 2)

str_replace_nth(x = "pigpig", pattern = "pig", replacement = "dog", n = 2)

str_replace_nth(x = "pigpig", pattern = "pig", replacement = "gy", n = 2)

}
