% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_right.R
\name{str_right}
\alias{str_right}
\title{Return the n number of characters from the right}
\usage{
str_right(text, n)
}
\arguments{
\item{text}{String character to pass through}

\item{n}{Integer specifying the number of characters to return from the right.}
}
\value{
Returns a formatted string.
}
\description{
This is a convenience wrapper function around \code{str_sub()}, and is analagous
to \code{str_left()}.
}
\examples{
str_right("Wittgenstein", 5)

}
