% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_right.R
\name{str_right}
\alias{str_right}
\title{Return the n number of characters from the right}
\usage{
str_right(x, n)
}
\arguments{
\item{x}{String character to pass through}

\item{n}{Integer specifying the number of characters to return from the right.}
}
\description{
Wrapper convenience function
}
